Remove unused code and move message creator interface implementation.
